% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texport.R
\name{texport}
\alias{texport}
\title{TeX Exporting Function}
\usage{
texport(x, file_name = "figure_1", operating_system = "macOS")
}
\arguments{
\item{x}{plot object to be exported to the current working directory}

\item{file_name}{name of resulting figure}

\item{operating_system}{specify the operating to ensure smooth handling of
'\' when the filepath for the new figure is created}
}
\value{
a TeX file containing the plot that is being exported
}
\description{
This function exports a plot object to a tex file using the awesome
tikzDevice engine.
}
\details{
Use \code{texport()} if you really need to insert your file in a LaTeX document
as a TeX code (possibly to do some changes that are easier to do in \LaTeX
than in R)
}
