% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{tsp}
\alias{tsp}
\title{Plot relative changes (percentage changes)}
\usage{
tsp(
  line_1_data = NULL,
  line_2_data = NULL,
  line_3_data = NULL,
  start_date,
  end_date,
  y_ticks = 100,
  labels_margin = 0,
  y_axis = TRUE,
  y_lim = c(-100, 100),
  mar_values = c(2, 3, 2, 1)
)
}
\arguments{
\item{mar_values}{}
}
\description{
'tsp()' takes as an input a vector containing time series and produces as an
output a plot showing graphically the relative (compared to initial value)
change of the time series over time.
}
\details{
It was moved from the \code{ter} package as a part of its \emph{conscious uncoupling}
into several smaller packages. Needs heavy editing because the body of \code{tsp()}
is a \emph{copy-paste} of \code{R} code which I wrote about a research project and
expects data frame in a very specific format (each input is a two-column
data frame that contains dates in the first column and values in the second
column). Furthermore, it is only limited to plotting three lines because it
uses \code{\link[=plot]{plot()}} under the hood (instead of the more suitable \code{\link[=matplot]{matplot()}}
function).
}
