% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr.R
\name{tsr}
\alias{tsr}
\title{Function for plotting unadjusted (raw) time series}
\usage{
tsr(x, xlab = NULL, main = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Function for plotting unadjusted (raw) time series
}
\examples{
econres::mGDP |>
dplyr::select(c(Year, `United Kingdom`, `United States`, China, India, Germany)) |>
dplyr::filter(Year >= 1000) |>
dplyr::arrange(Year) |>
tsr(xlab = `Year`)


}
