% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss.R
\name{tss}
\alias{tss}
\title{Plot stock prices}
\usage{
tss(
  start_date = "1983-06-27",
  end_date = Sys.Date(),
  stock_ticker_1 = "^GSPC",
  stock_ticker_2 = "BRK-A",
  label_1 = "S&P 500",
  label_2 = "Berkshire \\nHathaway",
  freq_data = "monthly",
  annotation_x = NULL,
  annotation_y = NULL,
  annotation_text = NULL
)
}
\arguments{
\item{start_date}{A date in the format "YYYY-MM-DD" (e.g. "1983-06-27")}

\item{end_date}{A date in the format "YYYY-MM-DD" (e.g. "1983-06-27")}

\item{stock_ticker_1}{Character to provide the ticker of the first stock}

\item{stock_ticker_2}{Character to provide the ticker of the second stock}

\item{freq_data}{Character to indicate the frequency of financial data:
'daily', 'weekly', 'monthly' (default), 'yearly'}

\item{annotation_x}{The x-axis coordinate(s) of one or more annotations}

\item{annotation_y}{The y-axis coordinate(s) of one or more annotations}

\item{annotation_text}{The text(s) for one or more annotations}

\item{S_and_P_500_benchmark}{Logical to indicate whether the S&P 500
index should be plotted as a benchmark, the default value is TRUE}
}
\description{
\code{tss()} plots up to two time series containing index(es) of stock prices
for a specified time period.
}
\details{
This function was moved from \code{ter} package as part of its
\emph{conscious uncoupling} into smaller and more specialized packages.
}
