% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr.R
\name{tsr}
\alias{tsr}
\title{Function for plotting unadjusted (raw) time series}
\usage{
tsr(x, currency = "$", unit = "", ...)
}
\arguments{
\item{x}{A dataframe (in wide format) where the first column is a vector of
years and the rest of the columns are vectors of values presenting time
series for one or more economic variables.}

\item{currency}{The currency in which the time series are quoted (defaulting
to $).}

\item{unit}{The unit in which the time series are measured (mlns, kgs, etc.)}

\item{...}{Arguments to be passed to methods of \code{\link[=plot]{plot()}} and \code{\link[=lines]{lines()}}, such
as \link{graphical parameters} (see \link{par}).}
}
\value{
A plot object.
}
\description{
\code{tsr()} takes a dataframe with raw time series data and uses the data to
create a time series plot with multiple lines for the time period which the
data covers.
}
\examples{
mGDP |>
dplyr::select(c(Year, `United Kingdom`, `United States`, China, India, Germany)) |>
dplyr::filter(Year >= 1000) |>
dplyr::arrange(Year) |>
tsr(xlab = 'Year')


}
